requirements.txt
fastapi==0.110.2
matplotlib==3.10.1
numpy==2.2.5
opencv-python==4.11.0.86
pillow==11.2.1
webcolors==24.11.1
uvicorn==0.34.2
python-multipart==0.0.20
pydantic_settings==2.0.0
